# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/resources/{id}": {
      "get": {
        "operationId": "ResourcesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Get resource by ID",
        "tags": [
          "Resources"
        ]
      },
      "patch": {
        "operationId": "ResourcesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resource updated"
          }
        },
        "summary": "Update resource",
        "tags": [
          "Resources"
        ]
      },
      "delete": {
        "operationId": "ResourcesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Resource deleted"
          }
        },
        "summary": "Delete resource",
        "tags": [
          "Resources"
        ]
      }
    },
...
    "/api/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Get event by ID",
        "tags": [
          "Events"
        ]
      },
      "patch": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated"
          }
        },
        "summary": "Update event",
        "tags": [
          "Events"
        ]
      },
      "delete": {
        "operationId": "EventsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted"
          }
        },
        "summary": "Delete event",
        "tags": [
          "Events"
        ]
      }
    },
...
    "/api/registrations/event/{eventId}": {
      "get": {
        "operationId": "RegistrationsController_findByEvent",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get registrations for an event",
        "tags": [
          "Registrations"
        ]
      }
    },
...
    "/api/attendance": {
      "get": {
        "operationId": "AttendanceController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all attendance records",
        "tags": [
          "Attendance"
        ]
      }
    },
    "/api/attendance/registration/{regId}": {
      "patch": {
        "operationId": "AttendanceController_updateByRegistration",
        "parameters": [
          {
            "name": "regId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAttendanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attendance updated"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Update attendance by registration ID",
        "tags": [
          "Attendance"
        ]
      }
    },
    "/api/event-staff": {
      "get": {
        "operationId": "EventStaffController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all staff assignments",
        "tags": [
          "Event Staff"
        ]
      },
      "post": {
        "operationId": "EventStaffController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventStaffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Assignment created"
          }
        },
        "summary": "Create staff assignment",
        "tags": [
          "Event Staff"
        ]
      }
    },
    "/api/event-staff/{id}": {
      "patch": {
        "operationId": "EventStaffController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventStaffDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assignment updated"
          }
        },
        "summary": "Update staff assignment",
        "tags": [
          "Event Staff"
        ]
      },
      "delete": {
        "operationId": "EventStaffController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment deleted"
          }
        },
        "summary": "Delete staff assignment",
        "tags": [
          "Event Staff"
        ]
      }
    },
    "/api/event-resources": {
      "get": {
        "operationId": "EventResourcesController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all resource allocations",
        "tags": [
          "Event Resources"
        ]
      },
      "post": {
        "operationId": "EventResourcesController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventResourceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Allocation created"
          }
        },
        "summary": "Create resource allocation",
        "tags": [
          "Event Resources"
        ]
      }
    },
    "/api/event-resources/{id}": {
      "patch": {
        "operationId": "EventResourcesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventResourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Allocation updated"
          }
        },
        "summary": "Update resource allocation",
        "tags": [
          "Event Resources"
        ]
      },
      "delete": {
        "operationId": "EventResourcesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Allocation deleted"
          }
        },
        "summary": "Delete resource allocation",
        "tags": [
          "Event Resources"
        ]
      }
    },
...
  },
  "info": {
    "title": "EMCP API",
    "description": "Event Management & Coordination Platform REST API. Role is passed via the x-user-role header. User ID is passed via the x-user-id header for user-specific operations.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UpdateResourceDto": {
        "type": "object",
        "properties": {
          "resourceName": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          }
        }
      },
...
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "eventDate": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          },
          "organizerId": {
            "type": "string"
          },
          "venueId": {
            "type": "string"
          }
        }
      },
...
      "UpdateAttendanceDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "PRESENT",
            "enum": [
              "PENDING",
              "PRESENT",
              "ABSENT"
            ]
          },
          "attendanceTime": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateEventStaffDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "example": "e1"
          },
          "staffId": {
            "type": "string",
            "example": "u5"
          },
          "role": {
            "type": "string",
            "example": "Main Stage Setup"
          },
          "shiftStart": {
            "type": "string",
            "example": "09:00"
          },
          "shiftEnd": {
            "type": "string",
            "example": "12:00"
          },
          "status": {
            "type": "string",
            "example": "ASSIGNED"
          }
        },
        "required": [
          "eventId",
          "staffId",
          "role",
          "shiftStart",
          "shiftEnd"
        ]
      },
      "UpdateEventStaffDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string"
          },
          "staffId": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "shiftStart": {
            "type": "string"
          },
          "shiftEnd": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "CreateEventResourceDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "example": "e1"
          },
          "resourceId": {
            "type": "string",
            "example": "r1"
          },
          "quantityUsed": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "eventId",
          "resourceId",
          "quantityUsed"
        ]
      },
      "UpdateEventResourceDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string"
          },
          "resourceId": {
            "type": "string"
          },
          "quantityUsed": {
            "type": "number"
          }
        }
      },
...
    }
  }
}